EIP1-12986: Fix type issues
