+++ src/main/kotlin/uk/gov/dluhc/notificationsapi/dto/ContactDetailsDto.kt
@@ -16,3 +16,20 @@ data class AddressDto(
     val area: String?,
     val postcode: String,
 )
+
+fun ContactDetailsDto.mapEroContactFields(personalisation: MutableMap<String, String>) {
+    personalisation["LAName"] = localAuthorityName
+    personalisation["eroPhone"] = phone
+    personalisation["eroWebsite"] = website
+    personalisation["eroEmail"] = email
+    with(address) {
+        personalisation["eroAddressLine1"] = getSafeValue(property)
+        personalisation["eroAddressLine2"] = street
+        personalisation["eroAddressLine3"] = getSafeValue(town)
+        personalisation["eroAddressLine4"] = getSafeValue(area)
+        personalisation["eroAddressLine5"] = getSafeValue(locality)
+        personalisation["eroPostcode"] = postcode
+    }
+}
+
+fun getSafeValue(input: String?): String = input ?: ""
