+++ src/main/kotlin/uk/gov/dluhc/notificationsapi/mapper/SignatureResubmissionTemplatePreviewDtoMapper.kt
@@ -2,9 +2,14 @@ package uk.gov.dluhc.notificationsapi.mapper
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
-import uk.gov.dluhc.notificationsapi.dto.*
+import uk.gov.dluhc.notificationsapi.dto.CommonTemplatePreviewDto
+import uk.gov.dluhc.notificationsapi.dto.GenerateSignatureResubmissionTemplatePreviewDto
+import uk.gov.dluhc.notificationsapi.dto.LanguageDto
+import uk.gov.dluhc.notificationsapi.dto.NotificationType
 import uk.gov.dluhc.notificationsapi.dto.NotificationType.SIGNATURE_RESUBMISSION
 import uk.gov.dluhc.notificationsapi.dto.NotificationType.SIGNATURE_RESUBMISSION_WITH_REASONS
+import uk.gov.dluhc.notificationsapi.dto.SignatureResubmissionPersonalisationDto
+import uk.gov.dluhc.notificationsapi.dto.mapEroContactFields
 import uk.gov.dluhc.notificationsapi.models.GenerateSignatureResubmissionTemplatePreviewRequest
 import uk.gov.dluhc.notificationsapi.models.SignatureResubmissionPersonalisation
 import uk.gov.dluhc.notificationsapi.models.SourceType
@@ -40,6 +45,8 @@ class SignatureResubmissionTemplatePreviewDtoMapper {
     ): Map<String, Any> {
         val personalisationMap = mutableMapOf<String, Any>()
 
+        val eroContactDetailsDto = request.personalisation.eroContactDetails.let(eroContactDetailsMapper::fromApiToDto)
+
         with(request.personalisation) {
             personalisationMap["applicationReference"] = applicationReference
             personalisationMap["firstName"] = firstName
@@ -47,17 +54,19 @@ class SignatureResubmissionTemplatePreviewDtoMapper {
             personalisationMap["rejectionReasons"] = mapSignatureRejectionReasons(commonTemplatePreviewDto.language, this)
             personalisationMap["rejectionFreeText"] = templatePersonalisationDtoMapper.getSafeValue(rejectionFreeText?.ifBlank { null })
             with(mutableMapOf<String, String>()) {
-                (eroContactDetails.let(eroContactDetailsMapper::fromApiToDto)).mapEroContactFields(this)
+                eroContactDetailsDto.mapEroContactFields(this)
                 personalisationMap.putAll(this)
             }
             personalisationMap["sourceType"] = sourceTypeMapper.toSourceTypeString(request.sourceType, commonTemplatePreviewDto.language)
 
-            personalisationMap["deadline"] = templatePersonalisationDtoMapper.getSafeValue(mapDeadline(
-                deadlineDate,
-                deadlineTime,
-                commonTemplatePreviewDto.language,
-                sourceTypeMapper.toFullSourceTypeString(request.sourceType, commonTemplatePreviewDto.language)
-            ))
+            personalisationMap["deadline"] = templatePersonalisationDtoMapper.getSafeValue(
+                mapDeadline(
+                    deadlineDate,
+                    deadlineTime,
+                    commonTemplatePreviewDto.language,
+                    sourceTypeMapper.toFullSourceTypeString(request.sourceType, commonTemplatePreviewDto.language),
+                ),
+            )
             personalisationMap["uploadSignatureLink"] = uploadSignatureLink
         }
